requirements.txt
# =============================================================================
# SPESION 3.0 - Requirements
# =============================================================================
# New in 3.0: ModelRouter, Evolving Memory (sqlite3 stdlib), Cognitive Loop,
# Heartbeat Runner, Context Compaction, Workspace Bootstrap.
# All new modules use Python stdlib + existing LangChain deps — no new packages.

# -----------------------------------------------------------------------------
# Core LangChain & LangGraph
# -----------------------------------------------------------------------------
langgraph>=0.2.0
langchain>=0.3.0
langchain-core>=0.3.0
langchain-community>=0.3.0
langchain-ollama>=0.2.0
langchain-openai>=0.2.0
langchain-anthropic>=0.2.0

# -----------------------------------------------------------------------------
# Vector Store & Embeddings
# -----------------------------------------------------------------------------
chromadb>=0.5.0
sentence-transformers>=3.0.0

# -----------------------------------------------------------------------------
# API Server (SPESION 2.0)
# -----------------------------------------------------------------------------
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
uvloop>=0.19.0

# -----------------------------------------------------------------------------
# Discord Bot (SPESION 2.0)
# -----------------------------------------------------------------------------
discord.py>=2.4.0

# -----------------------------------------------------------------------------
# Telegram Bot
# -----------------------------------------------------------------------------
python-telegram-bot>=21.0

# -----------------------------------------------------------------------------
# Voice Transcription (Whisper Local)
# -----------------------------------------------------------------------------
faster-whisper>=1.0.0

# -----------------------------------------------------------------------------
# Configuration & Validation
# -----------------------------------------------------------------------------
pydantic>=2.0.0
pydantic-settings>=2.0.0
python-dotenv>=1.0.0

# -----------------------------------------------------------------------------
# HTTP Client
# -----------------------------------------------------------------------------
httpx>=0.27.0

# -----------------------------------------------------------------------------
# External APIs
# -----------------------------------------------------------------------------
# ArXiv
arxiv>=2.1.0

# Finance
yfinance>=0.2.40

# Notion
notion-client>=2.2.0

# -----------------------------------------------------------------------------
# Document parsing
# -----------------------------------------------------------------------------
pypdf>=4.0.0

# -----------------------------------------------------------------------------
# Image OCR (screenshots)
# -----------------------------------------------------------------------------
pillow>=10.0.0
pytesseract>=0.3.10

# Exchanges (Bitget via ccxt)
ccxt>=4.4.0

# GitHub
PyGithub>=2.3.0

# Garmin
garminconnect>=0.2.0

# Google Calendar
google-api-python-client>=2.0.0
google-auth-httplib2>=0.2.0
google-auth-oauthlib>=1.2.0

# Web Search
duckduckgo-search>=6.0.0
tavily-python>=0.5.0

# -----------------------------------------------------------------------------
# Docker SDK (for code sandbox)
# -----------------------------------------------------------------------------
docker>=7.0.0

# -----------------------------------------------------------------------------
# Data Processing
# -----------------------------------------------------------------------------
pandas>=2.0.0
numpy>=1.24.0

# -----------------------------------------------------------------------------
# Utilities
# -----------------------------------------------------------------------------
python-dateutil>=2.8.0
tenacity>=8.0.0
APScheduler>=3.10.0
requests>=2.31.0

# -----------------------------------------------------------------------------
# Development (optional)
# -----------------------------------------------------------------------------
# pytest>=8.0.0
# pytest-asyncio>=0.23.0
# black>=24.0.0
# ruff>=0.4.0
# mypy>=1.10.0

